Move utils package to lib/internal
