+++ lib/cmd/init/init.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/Travix-International/Travix.Core.Adk/lib/cmd"
 	"github.com/Travix-International/Travix.Core.Adk/lib/context"
+	. "github.com/Travix-International/Travix.Core.Adk/lib/internal"
 	"github.com/Travix-International/Travix.Core.Adk/lib/scaffold"
-	"github.com/Travix-International/Travix.Core.Adk/utils/isEmptyPath"
 )
 
 type InitCommand struct {
@@ -37,8 +37,7 @@ func (cmd *InitCommand) Register(context context.Context) {
 
 			// First we'll check to see if the directory is empty. It's purely for safety purposes, to ensure we don't overwrite
 			// anyting special. The command line handling has already validated that the folder actually exists
-			isEmptyPath, err := isEmptyPath.IsEmptyPath(appPathAbsolute)
-			if !isEmptyPath || err != nil {
+			if empty, err := IsEmptyPath(appPathAbsolute); empty || err != nil {
 				log.Printf("The specified appPath '%s' does not appear to be an empty directory\n%v", appPathRelative, err)
 				return err
 			}
